chore(hentai): move hentai service codes into service package
