+++ sisterhood-hentai/src/commonMain/kotlin/sisterhood/hentai/service/KtorHentaiService.kt
@@ -1,4 +1,4 @@
-package sisterhood.hentai
+package sisterhood.hentai.service
 
 import sisterhood.Hentai
 import sisterhood.HentaiService
