+++ sisterhood-hentai/src/commonMain/kotlin/sisterhood/hentai/service/KtorHentaiServiceFactory.kt
@@ -1,4 +1,4 @@
-package sisterhood.hentai
+package sisterhood.hentai.service
 
 import io.ktor.client.*
 import io.ktor.client.engine.cio.*
