+++ sisterhood-hentai/src/commonMain/kotlin/sisterhood/hentai/service/ehentai/EHentaiProtocols.kt
@@ -1,4 +1,4 @@
-package sisterhood.hentai.ehentai
+package sisterhood.hentai.service.ehentai
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
