+++ sisterhood-hentai/src/commonMain/kotlin/sisterhood/hentai/service/hitomi/HitomiArtist.kt
@@ -1,10 +1,10 @@
-package sisterhood.hentai.hitomi
+package sisterhood.hentai.service.hitomi
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class HitomiCharacter(
-    @SerialName("character")
+data class HitomiArtist(
+    @SerialName("artist")
     val name: String
 )
