+++ sisterhood-hentai/src/commonMain/kotlin/sisterhood/hentai/service/hitomi/HitomiCharacter.kt
@@ -1,10 +1,10 @@
-package sisterhood.hentai.hitomi
+package sisterhood.hentai.service.hitomi
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class HitomiGroup(
-    @SerialName("group")
+data class HitomiCharacter(
+    @SerialName("character")
     val name: String
 )
