+++ sisterhood-hentai/src/commonMain/kotlin/sisterhood/hentai/service/hitomi/HitomiGroup.kt
@@ -1,10 +1,10 @@
-package sisterhood.hentai.hitomi
+package sisterhood.hentai.service.hitomi
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class HitomiParody(
-    @SerialName("parody")
+data class HitomiGroup(
+    @SerialName("group")
     val name: String
 )
