+++ sisterhood-hentai/src/commonMain/kotlin/sisterhood/hentai/service/hitomi/HitomiParody.kt
@@ -1,10 +1,10 @@
-package sisterhood.hentai.hitomi
+package sisterhood.hentai.service.hitomi
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class HitomiArtist(
-    @SerialName("artist")
+data class HitomiParody(
+    @SerialName("parody")
     val name: String
 )
