+++ sisterhood-hentai/src/commonMain/kotlin/sisterhood/hentai/service/hitomi/HitomiThumbnailSize.kt
@@ -1,4 +1,4 @@
-package sisterhood.hentai.hitomi
+package sisterhood.hentai.service.hitomi
 
 enum class HitomiThumbnailSize {
     SMALL, SMALLBIG, BIG;
